# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agmux API",
    "version": "current",
    "description": "Current agmux HTTP API surface used by UI and agents."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4821"
    }
  ],
  "paths": {
...
    "/api/ptys/{id}/input-history": {
      "put": {
        "summary": "Save input history for a PTY",
        "parameters": [
          { "$ref": "#/components/parameters/ptyId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Saved" }
        }
      }
    },
    "/api/ptys/{id}/kill": {
      "post": {
        "summary": "Kill a PTY",
        "parameters": [
          { "$ref": "#/components/parameters/ptyId" }
        ],
        "responses": {
          "200": { "description": "Killed" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/ptys/{id}/resume": {
      "post": {
        "summary": "Legacy runtime resume endpoint",
        "parameters": [
          { "$ref": "#/components/parameters/ptyId" }
        ],
        "responses": {
          "200": { "description": "Reused" },
          "410": { "description": "Deprecated resume path" }
        }
      }
    },
    "/api/ptys/{id}/task": {
      "post": {
        "summary": "Assign a task to a PTY",
        "parameters": [
          { "$ref": "#/components/parameters/ptyId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Assigned" },
          "400": { "description": "Bad request" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Clear PTY task assignment",
        "parameters": [
          { "$ref": "#/components/parameters/ptyId" }
        ],
        "responses": {
          "200": { "description": "Cleared" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/ptys/attach-tmux": {
      "post": {
        "summary": "Attach an existing tmux session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Attached" },
          "400": { "description": "Bad request" },
          "404": { "description": "Not found" },
          "409": { "description": "Conflict" }
        }
      }
    },
    "/api/ptys/launch": {
      "post": {
        "summary": "Launch a PTY from launch dialog options",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Launched" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/api/ptys/shell": {
      "post": {
        "summary": "Spawn shell PTY",
        "responses": {
          "200": { "description": "Spawned" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/api/readiness/trace": {
      "get": {
        "summary": "Read readiness decision trace",
        "parameters": [
          {
            "name": "ptyId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/readiness/report": {
      "post": {
        "summary": "Report explicit agent readiness from Claude/Codex callbacks",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Reported" },
          "400": { "description": "Bad request" },
          "404": { "description": "Running PTY not found" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ptyId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    }
  }
}